adding comments

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -7,11 +7,11 @@ public class StatsServiceTest {
     @Test
     public void ofAllSales() {
         StatsService cervice = new StatsService();
-        long[] monthlySalesReport = {
+        long[] monthlySalesReport = { // это массив продаж, может быть больше 2.5 милиарда
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int expectedAmount = 180;
-        long actualAmount = cervice.totalSalesAmount(monthlySalesReport);
+        long expectedAmount = 180; // здесь сумма может быть больше 2.5 милиарда
+        long actualAmount = cervice.totalSalesAmount(monthlySalesReport); // принимает сумму возможно > 2.5 милиарда
         Assertions.assertEquals(expectedAmount, actualAmount);
     }
 
@@ -19,21 +19,21 @@ public class StatsServiceTest {
     public void averageAmount() {
         StatsService service = new StatsService();
 
-        long[] monthlySalesReport = {
+        long[] monthlySalesReport = { // это массив продаж, может быть больше 2.5 милиарда
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int expectedMontArray = 15;
-        long actualMontArray = service.averageSales(monthlySalesReport);
+        long expectedMontArray = 15; // здесь сумма может быть больше 2.5 милиарда
+        long actualMontArray = service.averageSales(monthlySalesReport); // принимает сумму возможно > 2.5 милиарда
         Assertions.assertEquals(expectedMontArray, actualMontArray);
     }
 
     @Test
     public void maxAmount() {
         StatsService service = new StatsService();
-        long[] monthlySalesReport = {
+        long[] monthlySalesReport = { // это массив продаж, может быть больше 2.5 милиарда
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int expectedMaxSales = 5;
+        int expectedMaxSales = 5; // номер месяца < 2.5 милиарда
         long actualMaxales = service.maximumSaleDay(monthlySalesReport);
         Assertions.assertEquals(expectedMaxSales, actualMaxales);
     }
@@ -41,33 +41,33 @@ public class StatsServiceTest {
     @Test
     public void minAmount() {
         StatsService cervice = new StatsService();
-        long[] monthlySalesReport = {
+        long[] monthlySalesReport = { // это массив продаж, может быть больше 2.5 милиарда
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int expectedMinDay = 8;
-        long actualMinDay = cervice.minimumSalesDay(monthlySalesReport);
+        int expectedMinDay = 8; // номер месяца < 2.5 милиарда
+        int actualMinDay = cervice.minimumSalesDay(monthlySalesReport); // принимает номер месяца < 2.5 милиарда
         Assertions.assertEquals(expectedMinDay, actualMinDay);
     }
 
     @Test
     public void minAverageSales() {
         StatsService cervice = new StatsService();
-        long[] monthlySalesReport = {
+        long[] monthlySalesReport = { // это массив продаж, может быть больше 2.5 милиарда
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int expectedAveregeSales = 7;
-        long actualAveregeSales = cervice.minAverageSales(monthlySalesReport);
+        int expectedAveregeSales = 7; // номер месяца < 2.5 милиарда
+        int actualAveregeSales = cervice.minAverageSales(monthlySalesReport); // принимает номер месяца < 2.5 милиарда
         Assertions.assertEquals(expectedAveregeSales, actualAveregeSales);
     }
 
     @Test
     public void maxAveregeSales() {
         StatsService service = new StatsService();
-        long[] monthlySalesReport = {
+        long[] monthlySalesReport = { // это массив продаж, может быть больше 2.5 милиарда
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int expectedAveregeSales = 7;
-        long actualAveregeSales = service.maxAveregeSales(monthlySalesReport);
+        int expectedAveregeSales = 7; // номер месяца < 2.5 милиарда
+        int actualAveregeSales = service.maxAveregeSales(monthlySalesReport); // принимает номер месяца < 2.5 милиарда
         Assertions.assertEquals(expectedAveregeSales, actualAveregeSales);
     }
 }
